Separate structs JSONObject and JSONArray to each file
