+++ JSONParser/JSONParser/JSONValue.swift
@@ -21,45 +21,3 @@ protocol JSONDataForm {
     var totalCount: Int { get }
     func countType() -> [String: Int]
 }
-
-struct JSONObject: JSONDataForm {
-    private let jsonObject: [String: JSONValue]
-
-    init(jsonObject: [String: JSONValue]) {
-        self.jsonObject = jsonObject
-    }
-    
-    var typeName: typeName {
-        return .object
-    }
-    
-    var totalCount: Int {
-        return jsonObject.count
-    }
-    
-    func countType() -> [String : Int] {
-        return TypeInspector.countType(of: Array(jsonObject.values))
-    }
-}
-
-struct JSONArray: JSONDataForm {
-    private let jsonArray: [JSONValue]
-    
-    init(jsonArray: [JSONValue]) {
-        self.jsonArray = jsonArray
-    }
-    
-    var typeName: typeName {
-        return .array
-    }
-    
-    var totalCount: Int {
-        return jsonArray.count
-    }
- 
-    func countType() -> [String : Int] {
-        return TypeInspector.countType(of: jsonArray)
-    }
-}
-
-
